Modulo 2 - Reestruturando exemplos Laboratorio 1

+++ NovosLaboratorios/src/laboratorio1/exemplos/exemplo1/ContaCorrente.java
@@ -5,14 +5,14 @@ import java.util.List;
 public class ContaCorrente {
     private String nomeCorrentista;
     private Integer numeroConta;
-    private float saldoConta = 0;
+    private double saldoConta = 0;
 
     public ContaCorrente(String nomeCorrentista, Integer numeroConta) {
         this.nomeCorrentista = nomeCorrentista;
         this.numeroConta = numeroConta;
     }
 
-    public ContaCorrente(String titular, Integer nconta, float saldo){
+    public ContaCorrente(String titular, Integer nconta, double saldo){
         this.nomeCorrentista = titular;
         this.numeroConta = nconta;
         this.saldoConta = saldo;
@@ -26,15 +26,15 @@ public class ContaCorrente {
         return numeroConta;
     }
 
-    public float getSaldoConta() {
+    public double getSaldoConta() {
         return saldoConta;
     }
 
-    public void depositar(float valorDeposito){
+    public void depositar(double valorDeposito) {
         saldoConta = saldoConta + valorDeposito;
     }
 
-    public void sacar(float valorSaque){
+    public void sacar(double valorSaque){
         if (valorSaque <= saldoConta) {
             saldoConta = saldoConta - valorSaque;
         } else {
@@ -48,7 +48,7 @@ public class ContaCorrente {
         System.out.println("R$ " + saldoConta);
     }
 
-    public void realizarEmprestimo(float valor) {
+    public void realizarEmprestimo(double valor) {
         if(contaJaFoiNegativada(this)) {
             System.out.println("Emprestimo negado!");
         } else {
